Avoid eager download of index
Index are only downloaded on demand for hover or completion.
As a result, the results of such hover or completion may be delayed
(when index is necessary eg when local repo doesn't contain useful
data), but that avoid a lot of download and CPU consumption when just
reading poms or just hooking the language server without really using
it.

Signed-off-by: Mickael Istria <mistria@redhat.com>